virtio: widen the sector count before bounding a blk transfer
VioBlkTransfer's ASSERT(sectorCount * VIO_BLK_SECTOR_SIZE <= PAGE_SIZE)
multiplied a uint32_t by a small constant in 32 bits, so sectorCount ==
0x800000 gave a product of 0 and the bound passed for a transfer eight
million sectors long (docs/review-2026-07 §3). Latent -- every caller today
chunks at 8 sectors -- but this assert is the only bound the page-sized
bounce buffer has, so it has to hold for the caller that gets it wrong.

How verified: kmt's m6_blk suite (chunk-boundary, last-sector and
past-capacity transfers) and the FAT32 volume it all runs on; `make test`
green.

+++ drivers/virtio/blk.c
@@ -129,7 +129,12 @@ static NTSTATUS VioBlkTransfer(uint32_t type, uint64_t sectorLba, uint32_t secto
                                void *buffer)
 {
     ASSERT(VioBlkPresent);
-    ASSERT(sectorCount * VIO_BLK_SECTOR_SIZE <= PAGE_SIZE);
+    /* Widen before multiplying: sectorCount is uint32_t, so the product was
+     * computed in 32 bits and sectorCount == 0x800000 wrapped to 0 -- the
+     * bound passed for a transfer eight million sectors long. Latent (every
+     * caller chunks at 8 sectors) but the assert is the only bound the
+     * bounce buffer has. */
+    ASSERT((uint64_t)sectorCount * VIO_BLK_SECTOR_SIZE <= PAGE_SIZE);
     if (sectorLba + sectorCount > VioBlkCapacitySectors)
     {
         return STATUS_IO_DEVICE_ERROR;
